policy: Add attribution example (i.) regarding "tracing"
This is in response to case 13 where the jury voted to accept a work
where evidence of undeclared "tracing" was provided. I feel like the
jury most likely voted this way due to the evidence being confusing and
not due to a disagreement as to how to interpret the rules, but I would
rather clarify the rules to be on the safe side.